VectorStoreProject/requirements.txt
langchain==0.2.10
langchain-core==0.2.20
langchain-community==0.2.9
langchain-ollama==0.1.2
langchain-text-splitters==0.2.2
langserve==0.2.2
sse_starlette==2.1.2
fastapi==0.111.1
python-dotenv==1.0.1
httpx==0.27.0
